Restructure source and use cmake, add example test
