+++ .gitignore
@@ -1,0 +1,4 @@
+_deps/
+build/
+.vscode/
+Testing/
