+++ apps/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB SOURCES main.cpp)
+add_executable(crypto-platform-app ${SOURCES})
+target_link_libraries(crypto-platform-app crypto-platform)
