+++ crypto-platform/AutoTrader.cpp
