+++ crypto-platform/MerkelMain.cpp
@@ -1,8 +1,8 @@
-#include "MerkelMain.h"
-#include "OrderBookEntry.h"
-#include "CSVReader.h"
-#include "Wallet.h"
-#include "AutoTrader.h"
+#include <crypto-platform/MerkelMain.h>
+#include <crypto-platform/OrderBookEntry.h>
+#include <crypto-platform/CSVReader.h>
+#include <crypto-platform/Wallet.h>
+#include <crypto-platform/AutoTrader.h>
 #include <iostream>
 #include <vector>
 #include <string>
