+++ crypto-platform/MerkelMain.h
@@ -1,9 +1,9 @@
 #pragma once
 #include <vector>
-#include "OrderBookEntry.h"
-#include "OrderBook.h"
-#include "Wallet.h"
-#include "AutoTrader.h"
+#include <crypto-platform/OrderBookEntry.h>
+#include <crypto-platform/OrderBook.h>
+#include <crypto-platform/Wallet.h>
+#include <crypto-platform/AutoTrader.h>
 
 class MerkelMain
 {
