+++ crypto-platform/Wallet.cpp
@@ -1,5 +1,5 @@
-#include "Wallet.h"
-#include "CSVReader.h"
+#include <crypto-platform/Wallet.h>
+#include <crypto-platform/CSVReader.h>
 #include <iostream>
 
 Wallet::Wallet()
