+++ env.yaml
@@ -1,0 +1,7 @@
+name: crypto-platform
+channels:
+  - conda-forge
+  - nodefaults
+dependencies:
+  - cmake=3.26.3
+  - gtest=1.13.0
